test(stylesmixedin): rename tests
