+++ packages/tests/src/lib/tests.ts
@@ -2,4 +2,4 @@ export { fixtures as combinedFixtures } from './fixtures/combined'
 export { fixtures as elementFixtures } from './fixtures/elementTags'
 export { fixtures as textFixtures } from './fixtures/textTags'
 export { fixtures as styleObjectFixtures } from './fixtures/style-object'
-export { fixtures as stylesMixedInFixtures } from './fixtures/styles-mixed-in'
+export { fixtures as stylesMixedInFixtures } from './fixtures/styles-in-leaf'
